requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
Brotli==1.1.0
celery==5.4.0
celery-redbeat==2.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cssselect2==0.7.0
Django==5.1.1
django-allauth==64.2.0
django-allauth-ui==1.4.1
django-appconf==1.0.6
django-compressor==4.5.1
django-redbeat==0.3
django-timezone-field==7.0
django-unfold==0.40.0
django-widget-tweaks==1.5.0
djhtml==3.0.6
fonttools==4.54.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kombu==5.4.1
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.47
pycparser==2.22
pydyf==0.11.0
pyphen==0.16.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
rcssmin==1.1.2
redis==5.0.8
requests==2.32.3
rjsmin==1.2.2
RouterOS-api==0.18.1
six==1.16.0
slippers==0.6.2
sqlparse==0.5.1
tenacity==9.0.0
tinycss2==1.3.0
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
zopfli==0.2.3.post1
